Extract InternalVar handling into common base class
SimpleFBs and BasicFBs have internal vars. To clean-up the inheritence
hierarchy a common base class, as it is done in 4diac IDE, is
introduced.
