+++ src/core/basefb.h
@@ -1,0 +1,77 @@
+/*******************************************************************************
+ * Copyright (c) 2005, 2025 Profactor GmbH, ACIN, fortiss GmbH,
+ *                          Martin Erich Jobst, Johannes Kepler University Linz
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   Thomas Strasser, Gunnar Grabmair, Alois Zoitl, Gerhard Ebenhofer, Ingo Hegny
+ *                - initial implementation and rework communication infrastructure
+ *   Martin Jobst - account for data type size in FB initialization
+ *   Alois Zoitl  - exracted internal variable handling to new CBaseFB
+ *******************************************************************************/
+
+#pragma once
+
+#include "funcbloc.h"
+
+/*!\ingroup CORE
+ * \brief structure to hold the data needed for creating the internal vars
+ *
+ */
+struct SInternalVarsInformation {
+    TPortId mNumIntVars; //!< Number of internal vars
+    const CStringDictionary::TStringId *mIntVarsNames; //!< List of the internalvarsnames
+    const CStringDictionary::TStringId *mIntVarsDataTypeNames; //!< List of the data type names for the internal vars
+
+    TPortId getVarId(CStringDictionary::TStringId paInternalName) const;
+};
+
+class CBaseFB : public CFunctionBlock {
+  public:
+    ~CBaseFB() override = default;
+
+    CIEC_ANY *getVar(CStringDictionary::TStringId *paNameList, unsigned int paNameListSize) override;
+
+    int toString(char *paValue, size_t paBufferSize) const override;
+
+    size_t getToStringBufferSize() const override;
+
+#ifdef FORTE_TRACE_CTF
+    void traceInstanceData() override;
+#endif
+
+  protected:
+    CBaseFB(forte::core::CFBContainer &paContainer,
+            const SFBInterfaceSpec &paInterfaceSpec,
+            CStringDictionary::TStringId paInstanceNameId,
+            const SInternalVarsInformation *const paVarInternals);
+
+    /*! \brief Get the internal variable with given number
+     *
+     * Attention this function will not perform any range checks on the paVarIntNum parameter!
+     * @param paVarIntNum number of the internal variable starting with 0
+     * @return pointer to the internal variable
+     */
+    virtual CIEC_ANY *getVarInternal(size_t paVarIntNum) = 0;
+
+    const CIEC_ANY *getVarInternal(size_t paVarIntNum) const {
+      return const_cast<CBaseFB *>(this)->getVarInternal(paVarIntNum);
+    }
+
+    void setInitialValues() override = 0;
+
+  private:
+    /*!\brief Get the pointer to a internal variable of the basic FB.
+     *
+     * \param paInternalName StringId of the internal variable name.
+     * \return Pointer to the internal variable or 0.
+     */
+    CIEC_ANY *getInternalVar(CStringDictionary::TStringId paInternalName);
+
+    const SInternalVarsInformation *const cmVarInternals; //!< struct holding the information on the internal vars.
+};
